refactor: Move exit_status.dart to models folder
